+++ app/src/main/java/com/chouchongkeji/service/user/info/impl/AppPaymentInfoServiceImpl.java
@@ -412,7 +412,7 @@ public class AppPaymentInfoServiceImpl implements AppPaymentInfoService {
         /** 订单类型 **/
         payment.setType(orderType);
         /** 支付平台 **/
-        payment.setPayPlatform((byte) Constants.PAY_TYPE.ALI);
+        payment.setPayPlatform(Constants.PAY_TYPE.ALI);
         /** 卖家账号 **/
         payment.setSeller(aLiPayV2Vo.getSeller_email());
         /** 支付总价 **/
@@ -771,7 +771,7 @@ public class AppPaymentInfoServiceImpl implements AppPaymentInfoService {
         /** 订单类型 **/
         payment.setType(orderType);
         /** 支付平台 **/
-        payment.setPayPlatform((byte) Constants.PAY_TYPE.WX);
+        payment.setPayPlatform(Constants.PAY_TYPE.WX);
         /** 卖家账号 **/
         payment.setSeller(notifyData.getMch_id());
         /** 支付总价--分为单位 **/
@@ -816,7 +816,7 @@ public class AppPaymentInfoServiceImpl implements AppPaymentInfoService {
         /** 订单类型 **/
         payment.setType(orderType);
         /** 支付平台 **/
-        payment.setPayPlatform((byte) Constants.PAY_TYPE.yue);
+        payment.setPayPlatform(Constants.PAY_TYPE.yue);
         /** 卖家账号 **/
         payment.setSeller(sellUserId.toString());
         /** 支付总价--分为单位 **/
